rename matchers -> matcher, add NullMatch and PathSegment.
