+++ api/envoy/config/rbac/v2alpha/BUILD
@@ -9,7 +9,7 @@ api_proto_library(
     deps = [
         "//envoy/api/v2/core:address",
         "//envoy/api/v2/route",
-        "//envoy/type/matchers:metadata",
+        "//envoy/type/matcher:metadata",
     ],
 )
 
@@ -19,6 +19,6 @@ api_go_proto_library(
     deps = [
         "//envoy/api/v2/core:address_go_proto",
         "//envoy/api/v2/route:route_go_proto",
-        "//envoy/type/matchers:metadata_go_proto",
+        "//envoy/type/matcher:metadata_go_proto",
     ],
 )
