+++ api/envoy/config/rbac/v2alpha/rbac.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 import "validate/validate.proto";
 import "envoy/api/v2/core/address.proto";
 import "envoy/api/v2/route/route.proto";
-import "envoy/type/matchers/metadata.proto";
+import "envoy/type/matcher/metadata.proto";
 
 package envoy.config.rbac.v2alpha;
 option go_package = "v2alpha";
